Major Python client API rework, break down API & colossal class
