+++ trustgraph-base/trustgraph/api/exceptions.py
@@ -1,0 +1,6 @@
+
+class ProtocolException(Exception):
+    pass
+
+class ApplicationException(Exception):
+    pass
