+++ trustgraph-cli/scripts/tg-delete-flow-class
@@ -13,9 +13,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def delete_flow_class(url, class_name):
 
-    api = Api(url)
+    api = Api(url).flow()
 
-    class_names = api.flow_delete_class(class_name)
+    class_names = api.delete_class(class_name)
 
 def main():
