+++ trustgraph-cli/scripts/tg-get-flow-class
@@ -14,9 +14,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def get_flow_class(url, class_name):
 
-    api = Api(url)
+    api = Api(url).flow()
 
-    cls = api.flow_get_class(class_name)
+    cls = api.get_class(class_name)
 
     print(json.dumps(cls, indent=4))
