+++ trustgraph-cli/scripts/tg-graph-to-turtle
@@ -19,9 +19,9 @@ default_collection = 'default'
 
 def show_graph(url, flow_id, user, collection):
 
-    api = Api(url)
+    api = Api(url).flow().id(flow_id)
 
-    rows = api.flow(flow_id).triples_query(
+    rows = api.triples_query(
         s=None, p=None, o=None,
         user=user, collection=collection,
         limit=10_000)
