+++ trustgraph-cli/scripts/tg-invoke-llm
@@ -14,9 +14,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def query(url, flow_id, system, prompt):
 
-    api = Api(url)
+    api = Api(url).flow().id(flow_id).
 
-    resp = api.flow(flow_id).text_completion(system=system, prompt=prompt)
+    resp = api.text_completion(system=system, prompt=prompt)
 
     print(resp)
