+++ trustgraph-cli/scripts/tg-load-pdf
@@ -31,7 +31,7 @@ class Loader:
             metadata,
     ):
 
-        self.api = Api(url).flow(flow_id)
+        self.api = Api(url).flow().id(flow_id)
 
         self.user = user
         self.collection = collection
