+++ trustgraph-cli/scripts/tg-set-prompt
@@ -15,9 +15,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def set_system(url, system):
 
-    api = Api(url)
+    api = Api(url).config()
 
-    api.config_put([
+    api.put([
         ConfigValue(type="prompt", key="system", value=json.dumps(system))
     ])
