+++ trustgraph-cli/scripts/tg-set-token-costs
@@ -15,9 +15,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def set_costs(api_url, model, input_costs, output_costs):
 
-    api = Api(api_url)
+    api = Api(api_url).config()
 
-    api.config_put([
+    api.put([
         ConfigValue(
             type="token-costs", key=model,
             value=json.dumps({
