+++ trustgraph-cli/scripts/tg-show-config
@@ -13,9 +13,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def show_config(url):
 
-    api = Api(url)
+    api = Api(url).config()
 
-    config, version = api.config_all()
+    config, version = api.all()
 
     print("Version:", version)
     print(json.dumps(config, indent=4))
