+++ trustgraph-cli/scripts/tg-show-flow-classes
@@ -13,9 +13,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def show_flow_classes(url):
 
-    api = Api(url)
+    api = Api(url).flow()
 
-    class_names = api.flow_list_classes()
+    class_names = api.list_classes()
 
     if len(class_names) == 0:
         print("No flows.")
@@ -24,7 +24,7 @@ def show_flow_classes(url):
     classes = []
 
     for class_name in class_names:
-        cls = api.flow_get_class(class_name)
+        cls = api.get_class(class_name)
         classes.append((
             class_name,
             cls.get("description", ""),
