+++ trustgraph-cli/scripts/tg-show-flow-state
@@ -14,9 +14,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def dump_status(metrics_url, api_url, flow_id):
 
-    api = Api(api_url)
+    api = Api(api_url).flow()
 
-    flow = api.flow_get(flow_id)
+    flow = api.get(flow_id)
     class_name = flow["class-name"]
 
     print()
