+++ trustgraph-cli/scripts/tg-show-prompts
@@ -15,9 +15,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def show_config(url):
 
-    api = Api(url)
+    api = Api(url).config()
 
-    values = api.config_get([
+    values = api.get([
         ConfigKey(type="prompt", key="system"),
         ConfigKey(type="prompt", key="template-index")
     ])
@@ -25,7 +25,7 @@ def show_config(url):
     system = json.loads(values[0].value)
     ix = json.loads(values[1].value)
 
-    values = api.config_get([
+    values = api.get([
         ConfigKey(type="prompt", key=f"template.{v}")
         for v in ix
     ])
