+++ trustgraph-cli/scripts/tg-show-tools
@@ -15,15 +15,15 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def show_config(url):
 
-    api = Api(url)
+    api = Api(url).config()
 
-    values = api.config_get([
+    values = api.get([
         ConfigKey(type="agent", key="tool-index")
     ])
 
     ix = json.loads(values[0].value)
 
-    values = api.config_get([
+    values = api.get([
         ConfigKey(type="agent", key=f"tool.{v}")
         for v in ix
     ])
