+++ trustgraph-cli/scripts/tg-start-flow
@@ -13,9 +13,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def start_flow(url, class_name, flow_id, description):
 
-    api = Api(url)
+    api = Api(url).flow()
 
-    api.flow_start(
+    api.start(
         class_name = class_name,
         id = flow_id,
         description = description,
