+++ trustgraph-cli/scripts/tg-stop-flow
@@ -13,9 +13,9 @@ default_url = os.getenv("TRUSTGRAPH_URL", 'http://localhost:8088/')
 
 def stop_flow(url, flow_id):
 
-    api = Api(url)
+    api = Api(url).flow()
 
-    api.flow_stop(id = flow_id)
+    api.stop(id = flow_id)
 
 def main():
